% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkrace.R
\name{checkrace}
\alias{checkrace}
\title{Check whether a race was completed in a given year.}
\usage{
checkrace(race, year)
}
\arguments{
\item{race}{A character race identifier. See below for details.}

\item{year}{A number corresponding to requested race year with the format: \code{YYYY}.}
}
\value{
Error message if race was not conducted in requested year. Otherwise nothing.
}
\description{
This function is for internal use by the functions \code{\link{racedata}} and \code{\link{raceinfo}}
to ensure that the webscraping procedure does not return an error.
}
\details{
The \code{race} parameter accepts the following inputs:
\itemize{
\item \code{"tdf"} (Tour de France)
\item \code{"giro"} (Giro d'Italia)
\item \code{"vuelta"} (La Vuelta ciclista a España)
\item \code{"dauphine"} (Critérium du Dauphiné)
\item \code{"suisse"} (Tour de Suisse)
\item \code{"tirreno"} (Tirreno-Adriatico)
\item \code{"parisnice"} (Paris-Nice)
\item \code{"romandie"} (Tour de Romandie)
\item \code{"catalunya"} (Volta Ciclista a Catalunya)
\item \code{"pologne"} (Tour de Pologne)
\item \code{"basque"} (Itzulia Basque Country)
}
}
\examples{
checkrace(race = "tdf", year = 2020)
}
