% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raceinfo.R
\name{raceinfo}
\alias{raceinfo}
\title{Scrape race information from procyclingstats.com.}
\usage{
raceinfo(race, year)
}
\arguments{
\item{race}{A character race identifier. See below for details.}

\item{year}{A number corresponding to requested race year with the format: `YYYY`.}
}
\value{
a dataframe containing race information for `race` in `year` with the following
variables included: `stage`, `stage_id` and `stage_type`.
}
\description{
This function is for internal use by `cyclingscraper` and `racedata` to retrieve information about stages
of the requested race.
}
\details{
The `race` parameter accepts the following inputs: `"tdf"` (Tour de France),
`"grio"` (Giro d'Italia), `"vuelta"` (La Vuelta ciclista a España), `"dauphine"`
(Critérium du Dauphiné), `"suisse"` (Tour de Suisse), `"tirreno"` (Tirreno-Adriatico),
`"parisnice"` (Paris-Nice), `"romandie"` (Tour de Romandie), `"catalunya"` (Volta Ciclista a Catalunya),
`"pologne"` (Tour de Pologne), and `"basque"` (Itzulia Basque Country).
}
\examples{
raceinfo("tdf", 2022)
}
