% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racenames.R
\name{racenames}
\alias{racenames}
\title{Convert race identifier to full names.}
\usage{
racenames(race)
}
\arguments{
\item{race}{A character race identifier. See below for details.}
}
\value{
The full name of race as character.
}
\description{
This function is for internal use by \code{\link{racedata}} to (a) print progress
and (b) return correct full names of races in the column \code{race}.
in the data produced by the webscraping procedure.
}
\details{
The \code{race} parameter accepts the following inputs:
\itemize{
\item \code{"tdf"} (Tour de France)
\item \code{"giro"} (Giro d'Italia)
\item \code{"vuelta"} (La Vuelta ciclista a España)
\item \code{"dauphine"} (Critérium du Dauphiné)
\item \code{"suisse"} (Tour de Suisse)
\item \code{"tirreno"} (Tirreno-Adriatico)
\item \code{"parisnice"} (Paris-Nice)
\item \code{"romandie"} (Tour de Romandie)
\item \code{"catalunya"} (Volta Ciclista a Catalunya)
\item \code{"pologne"} (Tour de Pologne)
\item \code{"basque"} (Itzulia Basque Country)
}
}
\examples{
racenames("tdf")
}
