% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raceurlkey.R
\name{raceurlkey}
\alias{raceurlkey}
\title{Convert race identifiers to URL key}
\usage{
raceurlkey(race)
}
\arguments{
\item{race}{race identifier. See below for details.}
}
\value{
An URL key for procyclingstats.com corresponding to the race requested by user
}
\description{
This function is for internal use by `racedata` and `raceinfo` to convert
the race identifier into a URL to be passed into the HTML request of the webscraping procedure.
}
\details{
The `race` parameter accepts the following inputs: `"tdf"` (Tour de France),
`"grio"` (Giro d'Italia), `"vuelta"` (La Vuelta ciclista a España), `"dauphine"`
(Critérium du Dauphiné), `"suisse"` (Tour de Suisse), `"tirreno"` (Tirreno-Adriatico),
`"parisnice"` (Paris-Nice), `"romandie"` (Tour de Romandie), `"catalunya"` (Volta Ciclista a Catalunya),
`"pologne"` (Tour de Pologne) and `"basque"` (Itzulia Basque Country).
}
\examples{
raceurlkey("tdf")
}
