% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raceyears.R
\name{raceyears}
\alias{raceyears}
\title{List of years where races were completed.}
\usage{
raceyears()
}
\value{
A list containing race years for each race
}
\description{
This function is for internal use by \code{\link{cyclingscraper}} and \code{\link{checkrace}} to ensure
that HTML requests are only passed for years were a given race exists.
}
\examples{
raceyears()
}
