% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkrace.R
\name{checkrace}
\alias{checkrace}
\title{Check whether a race exists in a given year.}
\usage{
checkrace(race, year)
}
\arguments{
\item{race}{A character race identifier. See below for details.}

\item{year}{A number corresponding to requested race year with the format: `YYYY`.}
}
\value{
Error message if race was not conducted in requested year. Otherwise nothing.
}
\description{
This function is for internal use by the functions `racedata` and `raceoverview`
to ensure that the webscraping procedure does not return an error.
}
\details{
The `race` parameter accepts the following inputs: `"tdf"` (Tour de France),
`"grio"` (Giro d'Italia), `"vuelta"` (La Vuelta ciclista a España), `"dauphine"`
(Critérium du Dauphiné), `"suisse"` (Tour de Suisse), `"tirreno"` (Tirreno-Adriatico),
`"parisnice"` (Paris-Nice), `"romandie"` (Tour de Romandie), `"catalunya"` (Volta Ciclista a Catalunya),
`"pologne"` (Tour de Pologne) and `"basque"` (Itzulia Basque Country).
}
\examples{
checkrace(race = "tdf", year = 2020)
}
