% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racedata.R
\name{racedata}
\alias{racedata}
\title{Scrape stage-level race data from procyclingstats.com.}
\usage{
racedata(race, year, progress = TRUE, sleep = 1)
}
\arguments{
\item{race}{A character race identifier. See below for details.}

\item{year}{A number corresponding to requested race year with the format: `YYYY`.}

\item{progress}{Logical value indicating whether function should print progress. Default is `TRUE`.}

\item{sleep}{A number corresponding to seconds of sleep time between HTML requests
to ensure that servers are not overloaded. Minimum value is `1` (second) which is also the default.}
}
\value{
a dataframe containing stage-level data for `race` in `year` with the following
variables included: `race`, `year`, `stage`, `stage_id`, `stage_num`, `stage_type`, `date`,
`departure`, `arrival`, `parcours_type`, `distance`, `vertical_meters`, `profile_score`,
`startlist_quality`, `avg_speed_winner`, `won_how`, `win_type` and `km_solo`.
}
\description{
This function is for internal use by `cyclingscraper` to retrieve stage-level
data.
}
\details{
The `race` parameter accepts the following inputs: `"tdf"` (Tour de France),
`"grio"` (Giro d'Italia), `"vuelta"` (La Vuelta ciclista a España), `"dauphine"`
(Critérium du Dauphiné), `"suisse"` (Tour de Suisse), `"tirreno"` (Tirreno-Adriatico),
`"parisnice"` (Paris-Nice), `"romandie"` (Tour de Romandie), `"catalunya"` (Volta Ciclista a Catalunya),
`"pologne"` (Tour de Pologne), and `"basque"` (Itzulia Basque Country).
}
\examples{
racedata("tdf", 2022)
}
