% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raceinfo.R
\name{raceinfo}
\alias{raceinfo}
\title{Scrape race information from procyclingstats.com.}
\usage{
raceinfo(race, year)
}
\arguments{
\item{race}{A character race identifier. See below for details.}

\item{year}{A number corresponding to requested race year with the format: \code{YYYY}.}
}
\value{
a dataframe containing race information for \code{race} in \code{year} with the following
variables included: \code{stage}, \code{stage_id} and \code{stage_type}.
}
\description{
This function is for internal use by \code{\link{cyclingscraper}} and \code{\link{racedata}} to retrieve information about stages
of the requested race.
}
\details{
The \code{race} parameter accepts the following inputs:
\itemize{
\item \code{"tdf"} (Tour de France)
\item \code{"giro"} (Giro d'Italia)
\item \code{"vuelta"} (La Vuelta ciclista a España)
\item \code{"dauphine"} (Critérium du Dauphiné)
\item \code{"suisse"} (Tour de Suisse)
\item \code{"tirreno"} (Tirreno-Adriatico)
\item \code{"parisnice"} (Paris-Nice)
\item \code{"romandie"} (Tour de Romandie)
\item \code{"catalunya"} (Volta Ciclista a Catalunya)
\item \code{"pologne"} (Tour de Pologne)
\item \code{"basque"} (Itzulia Basque Country)
}
}
\examples{
raceinfo("tdf", 2022)
}
