% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racenames.R
\name{racenames}
\alias{racenames}
\title{Convert race identifier to full names.}
\usage{
racenames(race)
}
\arguments{
\item{race}{A character race identifier. See below for details.}
}
\value{
The full name of race as character.
}
\description{
This function is for internal use by `racedata` to (a) print progress
and (b) return correct full names of races in the column `race`.
in the data produced by the webscraping procedure.
}
\details{
The `race` parameter accepts the following inputs: `"tdf"` (Tour de France),
`"grio"` (Giro d'Italia), `"vuelta"` (La Vuelta ciclista a España), `"dauphine"`
(Critérium du Dauphiné), `"suisse"` (Tour de Suisse), `"tirreno"` (Tirreno-Adriatico),
`"parisnice"` (Paris-Nice), `"romandie"` (Tour de Romandie), `"catalunya"` (Volta Ciclista a Catalunya),
`"pologne"` (Tour de Pologne) and `"basque"` (Itzulia Basque Country).
}
\examples{
racenames("tdf")
}
