% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raceurlkey.R
\name{raceurlkey}
\alias{raceurlkey}
\title{Convert race identifiers to URL key}
\usage{
raceurlkey(race)
}
\arguments{
\item{race}{race identifier. See below for details.}
}
\value{
An URL key for procyclingstats.com corresponding to the race requested by user
}
\description{
This function is for internal use by \code{\link{racedata}} and \code{\link{raceinfo}} to convert
the race identifier into a URL to be passed into the HTML request of the webscraping procedure.
}
\details{
The \code{race} parameter accepts the following inputs:
\itemize{
\item \code{"tdf"} (Tour de France)
\item \code{"giro"} (Giro d'Italia)
\item \code{"vuelta"} (La Vuelta ciclista a España)
\item \code{"dauphine"} (Critérium du Dauphiné)
\item \code{"suisse"} (Tour de Suisse)
\item \code{"tirreno"} (Tirreno-Adriatico)
\item \code{"parisnice"} (Paris-Nice)
\item \code{"romandie"} (Tour de Romandie)
\item \code{"catalunya"} (Volta Ciclista a Catalunya)
\item \code{"pologne"} (Tour de Pologne)
\item \code{"basque"} (Itzulia Basque Country)
}
}
\examples{
raceurlkey("tdf")
}
